fix(chain-replay): only send txs when gateway in sorting state
If unlucky, it might happen that the gateway does the following:
- processes new payload messages
- process transaction A from S1, B from S2, C from S1 -> puts them in
the pool and not in the tof_snapshot because it is NOT sorting yet
- process fcu message and it is now sorting.

The tof_snapshot always maintain ordering, because it does push_front to
add txs. However, active orders in the mempool are different because
they group txs by sender. As such, when starting the first new frag,
what happens is that txs are copied from the active_txs, but will
processed on order A C B for the reason I just mentioned.